l10n: s/site.locale/site.active_lang/g
minimal-mistakes-jekyll does not support multiple languages. it means
`locale` in `_config.yml` is not what we need because `locale` is chosen
by users. we use polyglot for i18n.

polyglot provides site.active_lang, which is the language used. however,
the templates in minimal-mistakes-jekyll hard-code the `site.locale`.
`site.locale` must be replaced with `site.active_lang`.

to localize everything, you need to copy all the files that contains
`site.locale` from the gem. but for now, just replace `site.locale` in
what we need to override.